+++ src/elements/datetimeformat/datetimeformat-formatrange.ts
@@ -1,0 +1,30 @@
+import {html, nothing} from 'lit';
+
+import AbstractDateTimeFormatConsumer from './abstract-datetimeformat-consumer';
+
+export default class extends AbstractDateTimeFormatConsumer {
+  #value: string = '';
+
+  get value(): string {
+    return this.#value;
+  }
+
+  override render() {
+    if (this.start && this.end && this.providerElement) {
+      try {
+        this.#value =
+            this.providerElement.intlObject.formatRange(this.start, this.end);
+      } catch {}
+    }
+
+    return html`
+      <span role="none" part="value"
+        lang=${this.currentLang ?? nothing}
+        dir=${this.currentDir ?? nothing}
+      >${this.#value}</span>
+      <span aria-hidden="true" hidden>
+        <slot></slot>
+      </span>
+    `;
+  }
+}